backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
diffusers==0.27.2
transformers==4.40.0
peft==0.10.0
accelerate==0.29.3
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
scipy==1.13.0
supabase==2.4.6
google-generativeai==0.5.4
safetensors==0.4.3
datasets==2.19.0
omegaconf==2.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
numpy==1.26.4
tqdm==4.66.2
imagehash==4.3.1
requests==2.31.0
open-clip-torch==2.24.0
